requirements.txt
streamlit==1.11.0
plotly==5.9.0
pandas==1.4.3
Pillow==9.2.0